# mvn-entrypoint: substantial speed-up