… jdk-8-alpine/mvn-entrypoint.sh
… jdk-8-slim/mvn-entrypoint.sh
… jdk-8/mvn-entrypoint.sh
… jdk-9-slim/mvn-entrypoint.sh
… jdk-9/mvn-entrypoint.sh