Copying a reference tree with ca. 7600 files and ~410MiB
currently takes about 47 (!) seconds and thusly delays
container creation substantially.

Rewrite the copying script to address this:
1) by being less verbose in the log file and doing fewer
   parameter expansions, we shave off about 1 second
2) using cp's --reflink=auto we bring this down to about
   42 seconds, in addition to decreasing disk space used
   (where the underlying file system supports this)
3) using find's
   -exec <command> {} +
   syntax, rather than
   -exec <command> {} ;
   we're now at 32 seconds
4) using sh instead of bash, time goes down to about 22
   seconds
5) as a final optimisation, if the target directory is
   empty, we can just copy everything in one go, bringing
   the time down to under 2 seconds.

All of the above numbers are for the debian-slim version,
using alpine linux, step 4) takes about 15 seconds and
the time taken in step 5) doesn't change, as alpine's
busybox cp doesn't support --reflink=auto at the moment.